refactor(fbuild-build): split pipeline.rs into submodules to satisfy LOC gate
